chore(refactor): Move CEL code to conditions package
This is a residual task from #327 to re-organise the code
