+++ internal/conditions/types/type_adapter.go
@@ -1,7 +1,7 @@
 // Copyright 2021 Zenauth Ltd.
 // SPDX-License-Identifier: Apache-2.0
 
-package codegen
+package types
 
 import (
 	"encoding/json"
